m2/requirements.txt
# M4: trimodal inference + FastAPI. Install with: python3 -m pip install -r requirements.txt
# (or use .venv — see README)

# Core ML / media (m2_core.py)
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
librosa>=0.10.0
soundfile>=0.12.0
Pillow>=10.0.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Optional: Streamlit UI (app_main.py) — pip install -r requirements-streamlit.txt
